Fix for Play 2.0 ticket \#45 Google Closure Compiler require issues
Ticket url https://play.lighthouseapp.com/projects/82401/tickets/45-google-closure-compiler-require-issues

Fixed reverse ordering issue by applying distinct to the flattened dependency val, then taking the tail of the resulting list, reversing it, and prepending it to the head of the flattend dependency list. This results in the correct order seen in the workingIndex example of the test in the comments of the bug.

Removed the incorrect leftover require('something') statements in the compiled js source by applying the source tree require regex's replaceAllIn to both the full source and the minified source

Adding emacs autosave and ensime generated files to gitignore