Ticket75
Fixed by adding an unchanged (in map partial and flatmap partial) compileName val that stores the original name. Pattern-matched that name in the AssetsCompiler generated flatMap partial to discover if the compiler was compiling Coffeescript. If so, two compiler sequences are run, the passed compile function, and an additional closure compile. The initial coffescript is compiled to a temporary directory, the js added to that directory, and then the generated coffee js file is passed to the closure compiler. The resulting tuple is built and returned out of the match and used as normal.