Support routing bound to 'any' HTTP method. Related to Ticket #21
For those who want to do more fine grained handling of the request in
the controller then ideally we need a way to bind to any HTTP verb.
Like in Play 1.x with the support for '*'. The word 'any' is in
lowercase so as not to be confused with an actual HTTP verb. Related to
Ticket #21