Add OpenID realm parameter
The realm identifies the domain that the end user is being asked to
trust. Google/Gmail OpenID login uses it to determine the value of a
user's ID. So if your website uses different host names, the same Gmail
user will have different OpenIDs at each host — unless you specify the
realm parameter and set it to 'http://*.yourdomain.com'.