fixing scalax.io duplicate dependency issue related to Console
giter8 is pulling in and also depending on an older version of scalax.io (0.4.0) API. Having both 0.4.1 and 0.4.0 versions on the console's classpath was causing runtime exceptions when creating a new project with `play new`.

My first idea, excluding scalax.io from giter8, did not work unfortunately (since giter8 is relaying on the older scalax.io API).

So for now I just downgraded the scalax API in the Console to the previous version (0.4). We should be able to upgrade back to 0.4.1 once giter8 moves on to the new version.