Show database evolution errors when auto-applied
When using applyEvolutions.default=true and an error applying the evolution happens, that error is not logged.  The reason was because the evolution state was only checked before the evolution is attempted but not after.  Checking the evolution state after the evolution has been attempted results in the evolution error being logged.