Use scala-io-file v0.4.2 for Scala 2.10
This has been compiled against Scala 2.10.0 final (as has it's
dependency scala-arm v1.3) so the problems from PR #676 are finally laid
to rest :)

Happily, it's no longer necessary to add the javax.transaction
exclusion, as that's been dropped from the list of dependencies of
scala-io.

scala-io v0.4.2 was /only/ published against Scala v2.10 so far as I can
see, so I haven't updated the other build deps of scala-io to it, as
they are still running within SBT at Scala 2.9.2.