Update samples/java/zentasks/conf/initial-data.yml
Fix the initial-data.yml file as it has the "users:", "projects:", and "tasks:" top-level lists which cause the whole file to be loaded as a LinkedHashMap by Yaml which in-turn can't be read by Ebean because it doesn't have a manager that knows how to read-in LinkedHashMaps.  By removing these top-level lists and just leaving the data as a list then Yaml loads the whole data as a List which Ebean can then process correctly.  