Fix to avoid asynchronous Ebean fetchAhead (automatically fetches the next page)
When a page is accessed, Ebean LimitOffsetPagingQuery.java automatically
starts a new background thread to fetch the next page. This thread uses
a database connection that can be still alive after the HTTP request was
processed. On heavy load, this problem causes the following exception:
java.sql.SQLException: Timed out waiting for a free available
connection.

FetchAhead is useless in a stateless web application. Therefore, it must
be disabled using setFetchAhead(false).