Application jar should be the first in (dist/stage) generated start script
When using dist or stage commands a start script is generated. This starts
java with a classpath where the application jar is the last (the case for
dist) or it's part of a file expansion pattern (-cp "foo/*") which is often
evaluated alphabetically. Thus application provided classpath resources like
logger.xml, ehcache.xml or messages may or may not be used depending on other
jars that are part of the classpath.

This commit puts the application jar(s) as the first ones in the classpath,
so that resources provided by the application are definitely used.