Call #asJava recursively in play.api.Configuration.from
Now, for example, you need to write additionalConfiguration for FakeApplication like the following when it contains List[A]...

  FakeApplication(
    additionalConfiguration = Map("foo" -> List("a", "b", "c").asJava)
  )

This is a API for Scala, so it isn't cool to call #asJava in this way.
I changed Configuration.from(data: Map[String, Any]) a little,
and It now convert the value of Map[String, Any] internaly to fit with the API of typesafeconfig.