Show in the browser compilation errors only
By filtering SBT "problems" based on their severity, we can tell
Play to show in the browser compilation errors only.

Thanks to this fix, if the compiler indicates a warning followed
by an exception, the developer won't see the warning instead of
the exception in his browser.

Thanks to @baloo for his help.