Ebean @Encrypted annotation documentation
Information regarding how to set up the Ebean Encrypted annotation added to Play documentation in the Ebean section. If not properly set up, a very vague error will be presented to the user and could lead to a lot of confusion in the diagnosis. I added content that covers areas where the current Play and Ebean documentation do not.