Fixing #923 - cookie expiry is now further in the past
On a loaded server, it's possible that the cookie expiry time may
actually end up being in the future compared to the date header.  This
puts it one day in the past.  This also should be long enough to deal
with any issues regarding timezones and clocks out of sync.