Prefix form() with Form class in JavaGuide4.md
The ZenTask tutorial does not work if only `form()` is used. Java is not able to find the method since the `Controller` class from which `Application` is inherited does not provide such a method. To fix this, add the corresponding Class (located in `play.data.*`) as a prefix to the function call.