Switch to using ActorSystem.dispatchers.defaultGlobalDisptacher
Akka 2.2 changed the return type of ActorSystem.dispatcher from being
MessageDispatcher to being the less specific ExecutionContext
(ExecutionContext is a super type of MessageDispatcher).  This is the
only binary compatibility breaking change for Play, and so is the only
thing preventing users from upgrading to Akka 2.2.

The implementation of ActorSystem.dispatcher is
ActorSystem.dispatchers.defaultGlobalDispatcher.  The method signatures
of these methods haven't changed between 2.1 and 2.2, so be switching to
using them instead, we can allow Play users to upgrade to Akka 2.2 at
their own leisure.

Please note, this change does not upgrade Play to Akka 2.2, it only
makes Play binary compatible with both Akka 2.1 and Akka 2.2, so that
users can at their option upgrade to Akka 2.2.