Don't lose echo when initialising Play project settings
"play-interaction-mode" was recently introduced as a setting for Play
projects. The setting object was initialised when a Play project is
created. Unfortunately, initialising the setting object created a
JLine ConsoleReader object which, as a side-effect, disabled terminal
echo. Instead of creating a ConsoleReader at initialisation time, we
need to delay creation of the ConsoleReader until the setting object is
actually used, so echo can be restored afterwards.