Upgraded to fluentlenium 0.9.0
When I upgraded fluentlenium earlier, I accidentally only upgraded it
for the test dependency of the Play core project, but not the play-test
project. In fact, Play core should not have had a test dependency on
fluentlenium, since its tests don't use selenium at all.

So, I've got rid of the unnecessary test dependency (and added in
explicit dependencies on things that it brought in transitively that
were needed), and also upgraded play-test to fluentlenium 0.9.0.