Upgrade to Scala 2.10.3-RC3 and check iteratee memory leak is fixed
Scala 2.10.3 inclues a fix that allows memory to be freed when chaining
Future.flatMap calls (SI-7336). Iteratees make extensive use of
Future.flatMap, so upgrading to Scala 2.10.3 allows us to avoid memory
leaks when using iteratees (#1084). Memory leaks can occur when
processing large amounts of data or when working with connections that
can remain open for a long time.