Fix echo restore in run/start mode
Fixes #1815

Previously we were keeping track of what echo was before turning it off,
and then setting it back to that.  The problem with this is that jline
*always* turns echo off when creating a new console reader, and so it
we were always restoring it to off.  So now instead of worrying about
keeping track, we call the restore() method, which just tells jline to
set the terminal back to some reasonable defaults.