Initial Play tracing SPI
Included is a proposed tracing SPI for Play.  It is a reification of the
Play support for console does (with a number of fixes, improvements and
simplifications).  This work is on-going but is now in a form that is
ready for some initial review.

The primary tracing callback API is represented as Java interfaces and
can be found in the `play-instrumentation-spi` sub-project.

There is an example implementation of the SPI that computes average
statistics for requests.

[Updates]
* Removed extraneous printlns
* Added copyright notice
* Some explanatory comments
* Also a fix for recording the beginning and end of badRequest generation
* Ran formatting
* More formatting changes
* Fixed NPE opportunity
* Fixed "small" POST problems