Make sure specs tests are run by Specs2 not JUnit
When a Specs2 test has an @RunWith(classOf[JUnitRunner]) annotation then
it can be run by either JUnit or Specs2. But we want the tests to be run
by Specs2, so we tell junit-interface to ignore that particular runner.
Unfortunately, if SBT may ask JUnit to run the test. If it does so, then
the test will be executed but the result will be ignored by SBT.

This change is a hack to make sure that SBT always asks Specs2 to run the
test, never JUnit. We can make sure Specs2 is preferred by adding Specs2
as the last entry in the testFrameworks SBT setting. Since SBT runs
frameworks in reverse order, it will always run Specs2 before JUnit.