Allow to use a configured WebDriver when running with test browser.
Helpers.running and WithBrowser so far only accepted a WebDriver class
to instantiate the appropriate WebDriver.

This change allows to provide an already configured WebDriver instance.
This makes it possible to e.g. create an HtmlUnitDriver for a specific
browser version, like this:

    import org.openqa.selenium.htmlunit._
    import com.gargoylesoftware.htmlunit._
    val driver = new HtmlUnitDriver(BrowserVersion.FIREFOX_17)
    driver.setJavascriptEnabled(true)

    running(TestServer(3333), driver) { browser =>
      ..
    }

For the java version of WithBrowser it seems to be intended that for futher
customization of the browser `WithBrowser#provideBrowser` should be used:
There was no webDriver configuration possible, thus nothing is changed here.

Actually for me using the BrowserVersion.FIREFOX_17 fixed an issue
with jquery: integration tests failed with

    WebDriverException: com.gargoylesoftware.htmlunit.ScriptException: TypeError: Cannot find
    function addEventListener in object [object HTMLDocument]. (http://code.jquery.com/jquery-1.10.1.min.js#5)