Adds set repeated mapping
In the Forms there are list and seq to define a repeated mapping, but not set. Altought HTTP request doesn't permit to have a set, it can contains repetitions and the order is preserved, it is very common to have a data object having a member with Set semantic instead.