Incorporated sbt-web concerns into Play in order to reduce the repetition for sbt-web plugin configuration. Test coverage has been increased around the scripted tests for dist. The tests now also include testing that asset generation for less files happens correctly.
The documentation has been updated to reflect sbt-web usage. In addition some references to the now, defunct Play command have also been removed.

Finally the play start command has been changed so that it is now a convenience function for invoking stage and its associated binary. This was done because the sbt-web asset pipeline must run. However I think that start is now more reliable as a consequence.