Don't concatenate js files
Fixes #2192.

Let me start by saying I don't 100% understand our existing commonjs
support, what it's supposed to do etc.  This PR maintains the old
behaviour when the `commonJs` flag is supplied.  For some reason, the
commonjs support would pass in all siblings of a js file.  It would then
convert them to commonjs modules, namespacing them.  It would then
process all dependencies, and include all the dependencies.  For the
average project that didn't use commonjs, this meant that nothing got
included because there weren't any require statements.

But then we changed it to disable commonjs support by default, which
turned off this auto commonjs namespacing thing, and for some reason
that meant that the closure compiler just decided to concatenate all the
input files.

So, what this does, is when when commonJs is not enabled, it doesn't
pass all the siblings in as input files, it just passes the one input
file.  This stops the concatenation from happening.