Fix misleading tip about Enumeratee example
"events >>> Enumerator.eof &> toCometMessage" is inconsistent with the
code sample being referred to, and is incorrect since Enumerator.eof
shouldn't be sent through toCometMessage. I suspect that the code sample
originally read this way but was corrected, and the tip was not updated
along with it. Furthermore, "Enumerator.eof" is no longer necessary in
this section due to a fixed bug.