Fix misleading tip about Enumeratee example
"events >>> Enumerator.eof &> toCometMessage" is inconsistent with the code sample being referred to, and is incorrect since Enumerator.eof shouldn't be sent through toCometMessage. I suspect that the code sample originally read this way but was corrected, and the tip was not updated
along with it. I think it's just simpler to highlight the meaning of &> here since >>> is already covered in the preceding sample.

BTW I tried to test this change with "build run" as described in the documentation guideline but all it got me was "[error] (*:update) sbt.ResolveException: unresolved dependency: com.typesafe.play#sbt-plugin;2.3-SNAPSHOT: not found" and I did not know how to address this.