Refactored assets support
Assets are no longer added to anything that SBT considers to be on the
classpath, except in test.

In dev mode, they are given their own special non caching classloader.

In prod mode, they are packaged into their own artifact and added to the
distribution.

This change ensures that sub projects assets are correctly handled.

I also added a scripted test that tests specifically the sub project
scenarios in different modes.