Remove the scalaVersion setting from the Play sbt plugin
Play 2.3 will support multiple versions of Play so it's no longer
appropriate for the sbt plugin to provide a setting with a fixed
Scala version. The setting has been removed. Instead the Play
templates will provide an explicit scalaVersion setting instead.

Change all build file to accept the scala.version system property
to configure the version of Scala that they use. Most build files
default to Scala 2.10.4, but the build file that generates the
Play templates will generate build.sbt files with a default value
of 2.11.0.

With this commit, runtests works with Scala 2.10.4, but fails on
the documentation tests when run with Scala 2.11.0. I haven't
fixed the documentation tests as part of this commit because this
commit is blocking 2.3.