Deprecate and stop using the overflowingExecutionContext
Fixes #2947.

The overflowingExecutionContext is tricky to use safely so
this change removes all uses from Play. The object has been
deprecated for Play 2.3 and should be removed in Play 2.4.

In all cases the trampoline ExecutionContext should be used
instead since it is much safer and offers basically the same
performance.