Let exception thrown by delegate call propagate
If there are several action compositions and this one is not executed first (for example if declared at class level), then an expected exception thrown by the controller method (delegate.call) will create a RuntimeException even if the expected exception is controlled in another action composition. An exception thrown by authenticator can be caught here though.

Fix if formatting