Ensure appplication.conf log configuration works in prod
Fixes #1186.

The logger is configured in the constructor of Application.  However, it
was also being configured in the constructor of Server, this was so that
in dev mode, when the server first started up, before any application
existed, the default (or overridden default) configuration is applied.
The problem was, this was overriding the config done in Application,
since in prod mode, the server is instantiated after the application.

So, I moved the dev mode logging configuration concerns into
ReloadableApplication, which is instantiated during the instantiation of
the server in dev mode, and so effectively happens at the same time as
it used to.