Fix the bug that "play dependencies" command doesn't behave correctly when crossPaths := false.
When crossPaths is set to false, the names of the generated configuration reports have no version suffix, e.g "foo.xml" instead of "foo_2.10.xml". However, sbt tries to search for files with the scala version appended regardless of setting of crossPaths when computing dependencies. Therefore a FileNotFound error would be thrown.

I fix the issue by taking the value of crossPaths into consideration when looking for the reports to compute dependencies. If the crossPaths is set to false, we look for files without the version suffix, otherwise we look for files with the version suffix in their names.