Made Crypto an injectable component
Since this isn't a major component (the Play documentation doesn't
mention it) I didn't worry about breaking the Java API (the methods are
no longer static).

Also introduced a SimpleInjector implementation, that is map based, and
put Crypto in that for the default compile time dependency injection
components - since this is now the method that Crypto is looked up by,
for example when validating the session.