handle multipart fields with colon inside
if a form is submitted as multipart/form-data and a field in this form has a colon in its name
e.g. "message_id:signature", the original parameter name would be lost and translated to "message_idsignature"

This small change fixes this error. I'm also preparing a change for 2.4.x including a Specs-Testsuite.